Phase 4: any_value / memory-safety hardening
- 4a: delete both dead get_sample_at() methods (value::TimeSamples + TypedTimeSamples).
  The value::TimeSamples one returned const_cast<Sample*>(&*it) — a raw pointer into a
  reallocatable vector (dangling footgun). No callers remained after Phase 2c; the only
  user was unit-timesamples.cc, rewritten to has_sample_at() + get<float>() at the exact
  sample time (same intent, no raw pointer escape).
- 4b: value::TimeSamples::_data_offsets uint32_t -> size_t and BLOCKED_OFFSET UINT32_MAX ->
  SIZE_MAX, removing the 4GB byte-offset ceiling (the data buffer can exceed 4GB). Propagated
  through sort_flat_storage()'s param + temp, the get_data_offsets() return type, the byte_offset
  locals and the memory-usage sizeof. _array_counts stays uint32_t (it holds element counts).
- 4c: DEFINE_ROLE_TYPE_TRAIT now static_asserts sizeof/alignof(role) == sizeof/alignof(underlying)
  for every role type (auto-covers all ~23 + future ones). This makes the shared TimeSamples
  switch arms and the role path of any_value_raw_cast provably memory-safe. All role types pass.
- 4d: move any_value_raw_cast (unchecked force-cast) into value::detail so it cannot be reached
  without intent; the 8 internal callers (Value::as/get role branches) now qualify it. The
  checked any_value_cast stays public (it has external callers in crate-writer / c-tinyusd).

Builds clang + gcc; ctest 14/14; roundtrip 560 equivalent.

+++ src/timesamples.cc
@@ -43,7 +43,7 @@ inline void insertion_sort_samples(std::vector<value::TimeSamples::Sample>& samp
 inline void sort_flat_storage(
     std::vector<double>& times,
     Buffer<16>& blocked,
-    std::vector<uint32_t>& data_offsets,
+    std::vector<size_t>& data_offsets,
     std::vector<uint32_t>* array_counts) {
 
   const size_t n = times.size();
@@ -58,7 +58,7 @@ inline void sort_flat_storage(
   std::vector<double> sorted_times(n);
   Buffer<16> sorted_blocked;
   sorted_blocked.resize(n);
-  std::vector<uint32_t> sorted_offsets(data_offsets.size());
+  std::vector<size_t> sorted_offsets(data_offsets.size());
   std::vector<uint32_t> sorted_counts;
   if (array_counts && array_counts->size() == n) {
     sorted_counts.resize(n);
@@ -149,7 +149,7 @@ bool TimeSamples::reconstruct_binary_sample(size_t idx, Sample* sample) const {
       return true;
     }
 
-    const uint32_t byte_offset = _data_offsets[idx];
+    const size_t byte_offset = _data_offsets[idx];
     if (byte_offset == BLOCKED_OFFSET) {
       sample->blocked = true;
       return true;
@@ -392,7 +392,7 @@ size_t TimeSamples::estimate_memory_usage() const {
     total += _times.capacity() * sizeof(double);
     total += _blocked.capacity();
     total += _data.capacity();
-    total += _data_offsets.capacity() * sizeof(uint32_t);
+    total += _data_offsets.capacity() * sizeof(size_t);
     total += _array_counts.capacity() * sizeof(uint32_t);
 
     total += _samples.capacity() * sizeof(Sample);
@@ -409,7 +409,7 @@ size_t TimeSamples::estimate_actual_usage() const {
     total += _times.size() * sizeof(double);
     total += _blocked.size();  // vector<bool> counts elements, matches estimate pattern
     total += _data.size();
-    total += _data_offsets.size() * sizeof(uint32_t);
+    total += _data_offsets.size() * sizeof(size_t);
     total += _array_counts.size() * sizeof(uint32_t);
 
     total += _samples.size() * sizeof(Sample);
@@ -488,7 +488,7 @@ bool TimeSamples::get_scalar_impl(void *dst_, double t,
       };
       auto read_at = [&](size_t i, T *out) -> bool {
         if (i >= _data_offsets.size()) return false;
-        const uint32_t off = _data_offsets[i];
+        const size_t off = _data_offsets[i];
         if (off == BLOCKED_OFFSET) return false;
         if (static_cast<size_t>(off) + sizeof(T) > _data.size()) return false;
         std::memcpy(out, _data.data() + off, sizeof(T));
